fix: move login event handlers from app to login component
